requirements.txt
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.11.0
django-rest-swagger==0.3.7
djangorestframework==3.3.3
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
PyYAML==3.11
requests==2.10.0
wheel==0.24.0
six==1.10.0
stripe==1.35.0
